--- packages/autodesk-forge-nodejs-aws/public/css/main.css
@@ -1,23 +1,0 @@
-﻿html, body{
-  min-height: 100%;
-  height: 100%;
-}
-
-.fill{
-  height: calc(100vh - 100px);
-}
-
-body {
-  padding-top: 60px; /* space for the top nav bar */
-  margin-right: 30px;
-}
-
-#appBuckets {
-  overflow: auto;
-  width: 100%;
-  height: calc(100vh - 150px);
-}
-
-#forgeViewer {
-  width: 100%;
-}
