+++ packages/forge-viewmodels-nodejs-aws/package.json
