+++ packages/forge-viewmodels-nodejs-aws/routes/modelderivative.js
